DataB mysql hinzugefügt

+++ client/dataB.h
@@ -6,7 +6,7 @@
 void dataBInitDb(MYSQL* connection);
 MYSQL* dataBInit();
 void dataBExit(MYSQL* connection);
-void dataBStoreHum(MYSQL* connection, int nSensor1, double dHumidity1, int nSensor2, double dHumidity2, int pump);
+void dataBStoreHum(MYSQL* connection, int nSensor1, int dHumidity1, int nSensor2, int dHumidity2, int pump);
 double dataAll(MYSQL* connection);
 long dataBgetMeanHum(MYSQL* connection, int nStart, int nCount);
